[mohan]add: Maximum number in float

+++ FindMax/Generic.cs
@@ -9,7 +9,7 @@ namespace FindMaxNumber
     public class GenericMax
     {
 
-        public int FindMax(int first, int second, int third)
+        public float FindMax(float first, float second, float third)
         {
 
             if (first.CompareTo(second) > 0 && first.CompareTo(third) > 0)
